+++ CHANGELOG.md
@@ -6,8 +6,8 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- [#9](https://github.com/zendframework/zend-expressive-authorization-rbac/pull/9)
-  adds support for the 1.0.0-dev branch of zend-expressive-authorization.
+- [zendframework/zend-expressive-authorization-rbac#9](https://github.com/zendframework/zend-expressive-authorization-rbac/pull/9)
+  adds support for the 1.0.0-dev branch of mezzio-authorization.
 
 ### Changed
 
@@ -29,7 +29,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Added
 
-- Adds support for zend-expressive-authorization 0.2 and 0.3.
+- Adds support for mezzio-authorization 0.2 and 0.3.
 
 ### Changed